Change the data type of win percentage.

+++ Monty_Hall_Simulation_API/Monty_Hall_Simulation_API/Models/Result.cs
@@ -5,6 +5,6 @@
         public int GameCount { get; set; }
         public int WinCount { get; set; }
         public double WinRate { get; set; }
-        public int WinRatePercentage { get; set; }
+        public double WinRatePercentage { get; set; }
     }
 }
